adds stylesheet and changed update page ui

+++ fb_library/admin/mainpage.php
@@ -25,7 +25,12 @@
   <form action = "logout.php" method="POST"> 
     <div style="display: block; width: 100%;">
       <div class="logoutbtn-area">
-          <input type="submit" value="Logout" name="logout" class="logout-btn" />
+          <button type="submit" name="logout" class="logout-btn">
+            <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="white" class="bi bi-box-arrow-left" viewBox="0 0 16 16">
+              <path fill-rule="evenodd" d="M6 12.5a.5.5 0 0 0 .5.5h8a.5.5 0 0 0 .5-.5v-9a.5.5 0 0 0-.5-.5h-8a.5.5 0 0 0-.5.5v2a.5.5 0 0 1-1 0v-2A1.5 1.5 0 0 1 6.5 2h8A1.5 1.5 0 0 1 16 3.5v9a1.5 1.5 0 0 1-1.5 1.5h-8A1.5 1.5 0 0 1 5 12.5v-2a.5.5 0 0 1 1 0v2z"/>
+              <path fill-rule="evenodd" d="M.146 8.354a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L1.707 7.5H10.5a.5.5 0 0 1 0 1H1.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3z"/>
+            </svg>
+        </button>
       </div>
     </div>  
   </form>
@@ -38,10 +43,10 @@
       if (isset($_SESSION['login_user'])) {
         $userLoggedIn = $_SESSION['login_user'];
         $result = mysqli_query($con,"SELECT * FROM poll");
-
+          echo "<b><h1 style='margin-left:1.5em; margin-top:3%; font-size: 1.3em;'>Library&nbsp |&nbsp Feedback Entries</h1></b>";
           echo "<table border='1' id='respondents'>
           <tr>
-          <th>#</th>
+          <th style='width:5em;'>#</th>
           <th>Date</th>
           <th>Name</th>
           <th>Email Address</th>
@@ -79,7 +84,7 @@
                         Update
                       </button>
                     </a><br><br>
-                    <a href='delete-feedback.php?deleteid=".$id."'>
+                    <a href='delete-feedback.php?deleteid=".$id."' class='confirmation' >
                       <button class='delete-btn' />
                         Delete
                       </button>
@@ -93,6 +98,16 @@
       //header("Location: index.html");
       }
   ?>
+  
+  <script type="text/javascript">
+    var elems = document.getElementsByClassName('confirmation');
+    var confirmIt = function (e) {
+        if (!confirm('Are you sure that you want to delete this feedback?')) e.preventDefault();
+    };
+    for (var i = 0, l = elems.length; i < l; i++) {
+        elems[i].addEventListener('click', confirmIt, false);
+    }
+  </script>
 
   <script src='http://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js'></script>
